# backend/docs/openapi.apifox.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Light OSS API",
    "version": "1.0.0",
    "description": "OpenAPI 3.0 JSON for APIFOX import. Route coverage is derived from backend/internal/handler/router.go and excludes site content routes outside /api/v1 and /healthz."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local backend default"
    },
    {
      "url": "http://oss.underhear.cn",
      "description": "Gateway example"
    }
  ],
...
  "paths": {
...
    "/api/v1/buckets": {
      "get": {
        "tags": [
          "Buckets"
        ],
        "operationId": "listBuckets",
        "summary": "List buckets",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Bucket list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BucketListEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token. Error code: unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Failed to list buckets. Error code: bucket_list_failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Buckets"
        ],
        "operationId": "createBucket",
        "summary": "Create a bucket",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BucketCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bucket created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BucketEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Request body or bucket name is invalid. Error codes: invalid_request, invalid_bucket_name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token. Error code: unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Bucket already exists. Error code: bucket_exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Failed to create bucket. Error code: bucket_create_failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/buckets/{bucket}/folders/archive": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BucketParam"
        }
      ],
      "get": {
        "tags": [
          "Folders"
        ],
        "operationId": "downloadFolderArchive",
        "summary": "Download a folder as a ZIP archive",
        "description": "Streams a ZIP archive for the requested non-root folder path. The archive keeps the current folder name as the top-level directory, includes nested files, preserves empty subfolders created by the folder marker object, and excludes the internal .light-oss-folder marker file.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/FolderPathQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "ZIP archive stream.",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Bucket name or folder path is invalid. Error codes: invalid_bucket_name, invalid_folder_path, invalid_prefix.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token. Error code: unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Folder not found. Error code: folder_not_found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Folder archive preparation failed. Error codes: folder_lookup_failed, folder_archive_failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/buckets/{bucket}/entries": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BucketParam"
        }
      ],
      "get": {
        "tags": [
          "Folders"
        ],
        "operationId": "listExplorerEntries",
        "summary": "List direct entries under a folder prefix",
        "description": "Entries are ordered with directories first, then files, both sorted by name. next_cursor is an opaque cursor for the same prefix/search combination.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PrefixQuery"
          },
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/ExplorerLimitQuery"
          },
          {
            "$ref": "#/components/parameters/CursorQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Explorer entries page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplorerEntryListEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bucket name, prefix, or cursor is invalid. Error codes: invalid_bucket_name, invalid_prefix, invalid_cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token. Error code: unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Failed to list explorer entries. Error code: explorer_list_failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/buckets/{bucket}/objects": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BucketParam"
        }
      ],
      "get": {
        "tags": [
          "Objects"
        ],
        "operationId": "listObjects",
        "summary": "List objects",
        "description": "Objects are returned newest first. next_cursor is an opaque cursor for the same prefix filter.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PrefixQuery"
          },
          {
            "$ref": "#/components/parameters/ObjectListLimitQuery"
          },
          {
            "$ref": "#/components/parameters/CursorQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Object page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectListEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bucket name, prefix, or cursor is invalid. Error codes: invalid_bucket_name, invalid_prefix, invalid_cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token. Error code: unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Failed to list objects. Error code: object_list_failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
      "BucketParam": {
        "name": "bucket",
        "in": "path",
        "required": true,
        "description": "Bucket name. Must match ^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$.",
        "schema": {
          "type": "string",
          "pattern": "^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$",
          "minLength": 3,
          "maxLength": 63
        }
      },
...
      "FolderPathQuery": {
        "name": "path",
        "in": "query",
        "required": true,
        "description": "Folder path to delete. Must be a non-empty folder path ending with '/'.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 512
        }
      },
...
    },
...
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "request_id",
          "error"
        ],
        "properties": {
          "request_id": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "BucketCreateRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$",
            "minLength": 3,
            "maxLength": 63,
            "description": "Bucket name. Lowercase letters, digits, dot, and hyphen only."
          }
        }
      },
...
      "BucketEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "required": [
              "request_id",
              "data"
            ],
            "properties": {
              "request_id": {
                "type": "string"
              },
              "data": {
                "$ref": "#/components/schemas/Bucket"
              }
            }
          }
        ]
      },
      "BucketListEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "required": [
              "request_id",
              "data"
            ],
            "properties": {
              "request_id": {
                "type": "string"
              },
              "data": {
                "$ref": "#/components/schemas/BucketList"
              }
            }
          }
        ]
      },
...
    }
  }
}